# executor: tiny optimize import into from select performance